playground client module rename
